Fix wrong cast
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@1062 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplayfp/config.cpp
@@ -232,7 +232,7 @@ bool Player::sidCreate (sidbuilder *builder, SidConfig::model_t defaultModel,
             if ((i == 0) && !builder->getStatus())
                 return false;
             if (s)
-                s->sampling((long)cpuFreq, frequency, sampling, fastSampling);
+                s->sampling((float)cpuFreq, frequency, sampling, fastSampling);
             m_c64.setSid(i, s);
         }
     }
